Add folders for calculator and roborio
Calculator is an example that can be used to play around with.  Copied robot.py from last year to the roborio folder as a starting point.  Note that we will have to write the code over to be compliant with the rules.
Also updated the README document to explain the folders.